+++ extensions/word-search/src/types.ts
@@ -22,3 +22,7 @@ export enum WordType {
   "v" = "verb",
   "adj" = "adjective",
 }
+
+export interface Preferences {
+  defaultAction: string;
+}